Refine delegated auth integration boundaries
