+++ docs/.generated/plugin-sdk-api-baseline.sha256
@@ -1,2 +1,2 @@
-1f911e503e8ef6ffc76adb1201fcda139c0cf2287cff9c30ef237dd45c59e7f7  plugin-sdk-api-baseline.json
-aa8bfad04907ce8c1e604ee1c84e85f49ae885f610143ea422f8b41c18f38588  plugin-sdk-api-baseline.jsonl
+7591a9f3575a007b789bfcfaceb8a8ce4fcc05c7bd79396a3d179ee263237433  plugin-sdk-api-baseline.json
+99412365f32170370866f9aac233e56869cf1c3ae0584e9434b08a31a059ff35  plugin-sdk-api-baseline.jsonl
