+++ src/agents/openclaw-tools.ts
@@ -81,16 +81,13 @@ import { createUpdatePlanTool } from "./tools/update-plan-tool.js";
 import { createVideoGenerateTool } from "./tools/video-generate-tool.js";
 import { createWebFetchTool, createWebSearchTool } from "./tools/web-tools.js";
 import { resolveWorkspaceRoot } from "./workspace-dir.js";
-
 type OpenClawToolsDeps = {
   callGateway: typeof callGateway;
   config?: OpenClawConfig;
 };
-
 const defaultOpenClawToolsDeps: OpenClawToolsDeps = {
   callGateway,
 };
-
 let openClawToolsDeps: OpenClawToolsDeps = defaultOpenClawToolsDeps;
 
 /**
