+++ src/config/zod-schema.plugin-auth.ts
@@ -1,0 +1,16 @@
+import { z } from "zod";
+
+export const PluginEntryAuthSchema = z
+  .object({
+    delegatedAccess: z
+      .object({
+        enabled: z.boolean().optional(),
+        providers: z.array(z.string()).optional(),
+        audiences: z.array(z.string()).optional(),
+        scopes: z.array(z.string()).optional(),
+        chatTypes: z.array(z.enum(["direct", "group", "channel"])).optional(),
+      })
+      .strict()
+      .optional(),
+  })
+  .strict();
